refactor: cleanup and split folders
